Separate and organize root.tsx to multiple files
